libs: Add token-bucket library
The token bucket library implements the token bucket algorithm, to
be used for rate-limiting.

This commit implements basic token buckets, which contain tokens that
are refilled over time according to their refill parameter, up to a
maximum determined by the burst parameter.

Tokens can be consumed in a thread-safe way - consuming returns false
when there are not enough tokens available, and true when the operation
was successful.

Signed-off-by: Christian Pardillo Laursen <christian.pardillolaursen@citrix.com>